backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "langchain-community>=0.3.29",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.8",
    "pgvector>=0.4.1",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pypdf>=6.1.0",
    "sqlalchemy>=2.0.43",
    "supabase>=2.18.1",
    "torch>=2.8.0",
]
